Update tictactoe.php
Updated the game session so that it will automatically refresh the game board every second without refreshing the page. It calls itself, which returns a bunch of HTML/JS and rewrites the whole page with the HTML/JS code. We will need to update how we store the Win condition because it is not currently persistent. If someone can adjust the tictactoescripts.php file to store a WIN attribute to the DB then I can easily set the tictactoe.php to display that message instead of the board.