Update tictactoescripts.php
Works in conjunction with the recently updated tictactoe.php file to set Wins to be persistent without having to delete boards after a win. This version works both with the CEFNS site and Localhost. Probably. In order for this file to run properly, a WINNER attribute needs to be added to the BOARD table, with a default value of "No" and max length of 20.